#53: rotated example order
