+++ examples/example-1/.sequelizerc
