+++ examples/example-1/kafka-1-kafkajs/app.spec.ts
