+++ examples/example-1/postgres-1-sequelize/models/index.ts
