+++ examples/example-1/postgres-1-sequelize/seeders/20190101001337-demo-user.js
