+++ examples/example-1/postgres-2-knex/seeders/20180101001337-kill-bananas.js
