+++ examples/example-1/redis-1-ioredis/app.spec.ts
