+++ examples/example-2/package.json
@@ -7,23 +7,11 @@
     "test": "ts-node ./dockest.ts"
   },
   "dependencies": {
-    "dotenv": "^8.0.0",
-    "ioredis": "^4.14.0",
-    "kafkajs": "^1.10.0",
-    "knex": "^0.19.1",
-    "pg": "^7.12.0",
-    "pg-hstore": "^2.3.2",
-    "sequelize": "^5.12.3",
-    "sequelize-cli": "^5.4.0"
+    "ioredis": "^4.14.0"
   },
   "devDependencies": {
-    "@types/dotenv": "^6.1.0",
     "@types/ioredis": "^4.0.13",
     "@types/jest": "^24.0.17",
-    "@types/kafkajs": "^1.8.2",
-    "@types/knex": "^0.16.1",
-    "@types/pg": "^7.4.14",
-    "@types/sequelize": "^4.28.4",
     "jest": "^24.8.0",
     "ts-jest": "^24.0.2",
     "ts-node": "^8.0.2",
